requirements.txt
absl-py==2.0.0
app==0.0.1
APScheduler==3.10.4
asttokens==2.4.1
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
docutils==0.20.1
executing==2.0.1
ffpyplayer==4.5.0
Flask==2.3.3
Flask-APScheduler==1.12.4
flatbuffers==23.5.26
fonttools==4.43.1
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mediapipe==0.10.7
nest-asyncio==1.5.8
numpy==1.26.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
parso==0.8.3
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
sounddevice==0.4.6
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.6
wcwidth==0.2.9
Werkzeug==2.3.7
